requirements.txt
pyserial
matplotlib
pandas
numpy
scipy
